djangoapp/requirements.txt
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
iniconfig==2.0.0
packaging==24.1
pillow==11.0.0
pluggy==1.5.0
psycopg2-binary==2.9.10
ptw==1.0.1
PyJWT==1.7.1
pytest==8.3.3
pytest-django==4.9.0
rest-framework-simplejwt==0.0.2
sqlparse==0.5.1
